#61 Roles para Subject_User

+++ EsfotFeedback/app/Http/Controllers/SubjectUserController.php
@@ -13,19 +13,25 @@ class SubjectUserController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Subject_User::class);
         return new SubjectUserCollection(Subject_User::paginate(4));
     }
+
     public function show(Subject_User $subject_Users)
     {
+        $this->authorize('view', $subject_Users);
         return response()->json(new SubjectUserResource($subject_Users),200);
     }
+
     public function store(Request $request)
     {
         $subject_Users = Subject_User::create($request->all());
         return response()->json($subject_Users, 201);
     }
+
     public function delete(Subject_User $subject_Users)
     {
+        $this->authorize('delete', $subject_Users);
         $subject_Users->delete();
         return response()->json(null, 204);
     }
